Implementada extracao de parametros do método get
closed #9
