+++ rotas/query_string.py
@@ -1,0 +1,9 @@
+from flask import request
+
+from app import app
+
+
+@app.route('/name_lastname')
+def name_lastname():
+    name, lastname = request.args['name'], request.args['lastname']
+    return f'{name} {lastname}'
